Fix math

+++ lab_01/lab_01/lab_01/Converter.cs
@@ -43,7 +43,7 @@ namespace lab_01
 
         public PointF GetPointF(PointF point)
         {
-            return new PointF((int)GetX(point.X), (int)GetY(point.Y));
+            return new PointF(GetX(point.X), GetY(point.Y));
         }
 
         public PointF GetPointWithMargin(PointF point)
